GO2 can start iterations from supplied initial configuration
Main motivation was to study if GO only works because it starts
from decorana which is close to a solution. First test with varespec
imply that the same final solution can be reached from random
solutions, but not always, and the decorana solution is good one.

GOnormalize written long ago but never used was now edited to
scale the supplied configuration.