Add translation of Gaussian parameters to beta parameter
Other parameters are similar, but shape (alpha, gamma) varies
randomly.

The height is now equal, but it should be adjusted so that
Gaussian and beta have the same integral so that they predict
the same number of occurrences. Flat beta responses should be lower
and peaked should be higher than corresponding Gaussian.