beta parameters give same mass of presence as Gaussian models
The maximum height of beta response is selected so that the
response has the same mass (integrals) as the corresponding
Gaussian model. However, there is restriction that A0 <= 1, and
with the current parametrization peaked models (high alpha, gamma)
shoot over this limit, and when they are truncated, they have
lower mass and predict fewer presences than Gaussian models.

With random uniform shape the peaked responses are very common and
there are many species truncated to A0=1. Switching to log-uniform
could produce more often flat species.